requirements.txt
python==3.7
discord.py==1.7.2
